refactor(zero-cache): structure rm-vs-load benchmark
Move benchmark config, suite lifecycle, scenario execution, and consumer creation into named modules so the harness reads like maintained tooling instead of a one-off script.

Add unit coverage for config loading, scenario selection, and workload generation.
